http clientのテストを書く
