+++ lib/procon_bypass_man/background/jobs/concerns/has_server_pool.rb
@@ -1,49 +1,10 @@
 module ProconBypassMan::Background::HasServerPool
-  class ServerPool
-    def initialize(servers: )
-      if servers.nil? || servers.empty?
-        return
-      end
-
-      @servers = servers
-      if @servers.size >= 1
-        @index = 0
-      else
-        @index = nil
-      end
-    end
-
-    def pick
-      if @index.nil?
-        return @servers&.first
-      end
-      @servers[@index] or raise "bug!!!"
-    end
-    def server; pick; end
-
-    def next!
-      inc_index
-      if @servers[@index].nil?
-        reset
-        return
-      end
-    end
-
-    def reset
-      @index = 0
-    end
-
-    def inc_index
-      @index = @index + 1
-    end
-  end
-
   def reset_server_pool!
     @pool_server = nil
   end
 
   def pool_server
-    @pool_server ||= ServerPool.new(
+    @pool_server ||= ProconBypassMan::Background::ServerPool.new(
       servers: servers
     )
   end
